+++ testing/scenario_app/android/app/src/main/java/dev/flutter/scenarios/SurfacePlatformViewFactory.java
@@ -51,7 +51,7 @@ public final class SurfacePlatformViewFactory extends PlatformViewFactory {
   @SuppressWarnings("unchecked")
   @Override
   @NonNull
-  public PlatformView create(@NonNull Context context, int id, @Nullable Object args) {
+  public PlatformView create(@NonNull Context context, long id, @Nullable Object args) {
     if (preserveContext) {
       return new SurfacePlatformView(context);
     } else {
